LayoutTests/fast/history/link-inside-not-inside-any-expected.txt

:link inside :not inside :-webkit-any `:-webkit-any(:not(:link))`

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS done() became true
PASS successfullyParsed is true

TEST COMPLETE
Link should be green.

Link should be green.

Link should be red.

Link should be red.

Link should be green.

Link should be green.

Link should be green.

Link should be green.
